Push non-dependencies to top level
